refactor: extract shared handoff rendering into internal/handoff package
Eliminates duplicate handoff rendering between cmd/docket/handoff.go and
internal/mcp/tools_checkpoint.go by moving the logic into a shared
internal/handoff package. The cmd/docket wrapper delegates to the new
package, and the MCP checkpoint handler now calls handoff.WriteFile
directly instead of its own simplified copy.
